kbuild: minor improvement for thin archives build
The root built-in.o archive is currently generated before all object
files are built for the final link, due to final build of init/ after
version update. In practice it seems like it doesn't matter because
the archive symbol table does not change, but it is more logical to
create the final archive as the last step.

Change-Id: Idab18857b5b526f098752bf46bd3a742e7d4cd36
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Michal Marek <mmarek@suse.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>